qdrant-native-app-spcs/service/qdrant_driver/usecase_requirements.txt
jupyterlab
virtualenv
snowflake-snowpark-python[pandas]
snowflake-ml-python
matplotlib
pandas
sentence_transformers
tqdm
geopy
scikit-learn
qdrant-client[fastembed-gpu]==1.10.1
jupyterlab_nvdashboard
torch
accelerate==0.31.0
fastapi 
gunicorn 
uvicorn[standard]
pydantic==2.8.2
transformers
datasets
onnxruntime-gpu
ray[data]==2.33.0
ray[default]==2.33.0
ipywidgets